+++ classes/class.ilNotifications4PluginsUIHookGUI.php
@@ -1,16 +1,15 @@
 <?php
 
 use srag\DIC\Notifications4Plugins\DICTrait;
-use srag\Plugins\Notifications4Plugins\Utils\Notifications4PluginsTrait;
 
 /**
  * Class ilNotifications4PluginsUIHookGUI
  *
+ * @author studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  * @author Stefan Wanzenried <sw@studer-raimann.ch>
  */
 class ilNotifications4PluginsUIHookGUI extends ilUIHookPluginGUI {
 
 	use DICTrait;
-	use Notifications4PluginsTrait;
 	const PLUGIN_CLASS_NAME = ilNotifications4PluginsPlugin::class;
 }