+++ plugin.php
@@ -3,8 +3,8 @@
 require_once __DIR__ . "/vendor/srag/dic/src/PHPVersionChecker.php";
 
 $id = "notifications4pl";
-$version = "3.1.0";
-$ilias_min_version = "5.3.0";
+$version = "4.0.0";
+$ilias_min_version = "5.3.14";
 $ilias_max_version = "5.4.999";
 $responsible = "studer + raimann ag - Team Custom 1";
 $responsible_mail = "support-custom1@studer-raimann.ch";